Move all activities to a separate package
