+++ app/src/main/res/layout/activity_profile.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ProfileActivity">
+        tools:context=".activity.ProfileActivity">
     <TextView
             android:id="@+id/profileCreationLabel"
             android:layout_width="wrap_content"
